#148-MapBoardViewModel + MapViewModel 통합.
