+++ app/src/main/res/layout/map_board_layout.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
 
         <variable
             name="viewmodel"
-            type="com.teamdonut.eatto.ui.map.bottomsheet.MapBoardViewModel" />
+            type="com.teamdonut.eatto.ui.map.MapViewModel" />
 
     </data>
