+++ app/src/main/res/layout/map_fragment.xml
@@ -9,11 +9,6 @@
         <variable
             name="viewmodel"
             type="com.teamdonut.eatto.ui.map.MapViewModel" />
-
-        <variable
-            name="bottomsheetviewmodel"
-            type="com.teamdonut.eatto.ui.map.bottomsheet.MapBoardViewModel" />
-
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
@@ -121,7 +116,7 @@
             app:behavior_hideable="false"
             app:behavior_peekHeight="55dp"
             app:layout_behavior="@string/bottom_sheet_behavior"
-            app:viewmodel="@{bottomsheetviewmodel}" />
+            app:viewmodel="@{viewmodel}" />
 
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
 </layout>
